+++ public/stories/tic-tac-toe/styles/Content.module.scss
@@ -41,7 +41,7 @@
         display: inline-block;
         margin-right: sizing.$gutter;
         background: colors.$button;
-        color: colors.$cool;
+        color: colors.$very-light;
         border: 0;
         box-shadow: none;
         border-radius: 0;
